feat: simplly code
